+++ src/components/Calendar/CalendarRange/types.ts
@@ -4,10 +4,3 @@ export type DateRange = {
   startDate: Dayjs;
   endDate: Dayjs;
 };
-
-export const ClickState = {
-  START: "start",
-  END: "end",
-};
-
-export type ClickStateType = (typeof ClickState)[keyof typeof ClickState];
